fixes naming section
